requirements.txt
# torch==2.0.1+cu117
# pip install torch==2.0.1+cu117 --find-links https://download.pytorch.org/whl/cu117/torch/index.html

# torchvision==0.15.2+cu117
# pip install torchvision==0.15.2+cu117 --find-links https://download.pytorch.org/whl/cu117/torchvision/index.html

Pillow==10.0.1
transformers==4.49.0
numpy==1.26.4

pyside6==6.8.2.1
qasync==0.27.1
